Added configuration to contain titles and sizes
In order to be able to display the titles and sizes of the works, as
well as more fine-grained control over the order of the images. The
configuration was improved. Titles and sizes have been added and the
same configuration is now used for the navigation. Reading the folder
contents is no longer needed.